+++ packages/plugin-react-native-span-access/ios/BugsnagReactNativeAppStartPlugin.h
@@ -1,9 +1,10 @@
 #import <Foundation/Foundation.h>
 #import <BugsnagPerformance/BugsnagPerformancePlugin.h>
+#import <BugsnagReactNativePerformance/BugsnagPerformanceAppStartProvider.h>
 
 NS_ASSUME_NONNULL_BEGIN
 
-@interface BugsnagReactNativeAppStartPlugin: NSObject <BugsnagPerformancePlugin>
+@interface BugsnagReactNativeAppStartPlugin: NSObject <BugsnagPerformancePlugin, BugsnagPerformanceAppStartProvider>
 
 - (instancetype)init;
 - (instancetype)initWithTimeout:(NSTimeInterval)timeout;